Replaced old `SearchField` with the new one
